simplify explorer/navbar
