+++ packages/web-app/src/pages/explore.tsx
@@ -10,65 +10,49 @@ import Hero from 'containers/hero';
 import {Finance} from 'utils/paths';
 import Carousel from 'containers/carousel';
 import {Layout} from '../app';
-import {useWallet} from 'hooks/useWallet';
-import {useGlobalModalContext} from 'context/globalModals';
 
 const Explore: React.FC = () => {
-  const {isConnected, methods} = useWallet();
-  const {open} = useGlobalModalContext();
   const navigate = useNavigate();
 
-  // TODO move this into the navbar
-  const handleWalletButtonClick = () => {
-    if (isConnected) {
-      open('wallet');
-      return;
-    }
-    methods.selectWallet().catch((err: Error) => {
-      // To be implemented: maybe add an error message when
-      // the error is different from closing the window
-      console.error(err);
-    });
-  };
   return (
     <>
       <Container>
-        <ExploreNav onWalletClick={handleWalletButtonClick} />
+        <ExploreNav />
         <Hero />
         <Layout>
           <ContentWrapper>
             <Carousel />
+            <div className="h-20"></div>
+            <div className="p-2 m-5 space-y-1 bg-primary-100">
+              <p>
+                This is a temporarily added section for demonstration purposes.
+                It allows you to navigate to a mock dao to test daos URLs.
+              </p>
+              <ActionListItem
+                title={'Dao: 0x1234'}
+                subtitle={'ethereum mainnet'}
+                icon={<IconExpand />}
+                background={'white'}
+                onClick={() =>
+                  navigate(
+                    generatePath(Finance, {network: 'ethereum', dao: '0x1234'})
+                  )
+                }
+              />
+              <ActionListItem
+                title={'Dao: 0x1234'}
+                subtitle={'Rinkeby testnet'}
+                icon={<IconExpand />}
+                background={'white'}
+                onClick={() =>
+                  navigate(
+                    generatePath(Finance, {network: 'rinkeby', dao: '0x1234'})
+                  )
+                }
+              />
+            </div>
           </ContentWrapper>
         </Layout>
-        <div className="h-20"></div>
-        <div className="p-2 m-5 space-y-1 bg-primary-100">
-          <p>
-            This is a temporarily added section for demonstration purposes. It
-            allows you to navigate to a mock dao to test daos URLs.
-          </p>
-          <ActionListItem
-            title={'Dao: 0x1234'}
-            subtitle={'ethereum mainnet'}
-            icon={<IconExpand />}
-            background={'white'}
-            onClick={() =>
-              navigate(
-                generatePath(Finance, {network: 'ethereum', dao: '0x1234'})
-              )
-            }
-          />
-          <ActionListItem
-            title={'Dao: 0x1234'}
-            subtitle={'Rinkeby testnet'}
-            icon={<IconExpand />}
-            background={'white'}
-            onClick={() =>
-              navigate(
-                generatePath(Finance, {network: 'rinkeby', dao: '0x1234'})
-              )
-            }
-          />
-        </div>
         <div className="h-96"></div>
         <Footer />
       </Container>
